// src/schemas/ctrf/ctrf.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface CtrfSchema {
  reportFormat: ReportFormat;
  specVersion: string;
  reportId?: string;
  timestamp?: string;
  generatedBy?: string;
  results: {
    tool: {
      name: string;
      version?: string;
      extra?: {
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    summary: {
      tests: number;
      passed: number;
      failed: number;
      skipped: number;
      pending: number;
      other: number;
      suites?: number;
      start: number;
      stop: number;
      extra?: {
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    tests: {
      name: string;
      status: Status;
      duration: number;
      start?: number;
      stop?: number;
      suite?: string;
      message?: string;
      trace?: string;
      ai?: string;
      line?: number;
      rawStatus?: string;
      tags?: string[];
      type?: string;
      filePath?: string;
      retries?: number;
      flaky?: boolean;
      stdout?: string[];
      stderr?: string[];
      threadId?: string;
      browser?: string;
      device?: string;
      screenshot?: string;
      attachments?: {
        name: string;
        contentType: string;
        path: string;
        extra?: {
          [k: string]: unknown;
        };
        [k: string]: unknown;
      }[];
      parameters?: {
        [k: string]: unknown;
      };
      steps?: {
        name: string;
        status: Status;
        extra?: {
          [k: string]: unknown;
        };
        [k: string]: unknown;
      }[];
      extra?: {
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }[];
    environment?: {
      reportName?: string;
      appName?: string;
      appVersion?: string;
      buildName?: string;
      buildNumber?: string;
      buildUrl?: string;
      repositoryName?: string;
      repositoryUrl?: string;
      commit?: string;
      branchName?: string;
      osPlatform?: string;
      osRelease?: string;
      osVersion?: string;
      testEnvironment?: string;
      extra?: {
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    extra?: {
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  extra?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}

export const enum ReportFormat {
  CTRF = 'CTRF',
}
export const enum Status {
  passed = 'passed',
  failed = 'failed',
  skipped = 'skipped',
  pending = 'pending',
  other = 'other',
}